requirements.txt
Django>=5,<6
djangorestframework>=3.16,<4
Pillow>=10,<11
psycopg>=3.3,<4
